#34: Implement poor model loader

+++ hugame-desktop-window/src/main/java/dev/hugame/window/DesktopWindow.java
@@ -118,7 +118,7 @@ public class DesktopWindow implements Window {
 	}
 	
 	@Override
-	public void clear(int r, int g, int b) {
+	public void clear(float r, float g, float b) {
 		glClearColor(r, g, b, 1);	
 		glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 	}
